// src/gql/__generated__/graphql.ts
export type Page = ContentNode & DatabaseIdentifier & HierarchicalContentNode & HierarchicalNode & MenuItemLinkable & Node & NodeWithAuthor & NodeWithComments & NodeWithContentEditor & NodeWithFeaturedImage & NodeWithPageAttributes & NodeWithRevisions & NodeWithTemplate & NodeWithTitle & Previewable & UniformResourceIdentifiable & WithAcfPageBuilder & {
  __typename?: 'Page';
// ...
  pageBuilder?: Maybe<PageBuilder>;
// ...
};
// ...
export type PageBuilder = AcfFieldGroup & AcfFieldGroupFields & PageBuilder_Fields & {
  __typename?: 'PageBuilder';
  /**
   * The name of the field group
   * @deprecated Use __typename instead
   */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
  /** Field of the &quot;flexible_content&quot; Field Type added to the schema as part of the &quot;PageBuilder&quot; Field Group */
  sections?: Maybe<Array<Maybe<PageBuilderSections_Layout>>>;
};

/** The &quot;PageBuilderSectionsBeforefooterctaLayout&quot; Field Group. Added to the Schema by &quot;WPGraphQL for ACF&quot;. */
export type PageBuilderSectionsBeforefooterctaLayout = AcfFieldGroup & AcfFieldGroupFields & PageBuilderSectionsBeforefooterctaLayout_Fields & PageBuilderSections_Layout & {
  __typename?: 'PageBuilderSectionsBeforefooterctaLayout';
  /** Field of the &quot;textarea&quot; Field Type added to the schema as part of the &quot;PageBuilderSectionsBeforefooterctaLayout&quot; Field Group */
  description?: Maybe<Scalars['String']['output']>;
  /**
   * The name of the field group
   * @deprecated Use __typename instead
   */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
  /** Field of the &quot;text&quot; Field Type added to the schema as part of the &quot;PageBuilderSectionsBeforefooterctaLayout&quot; Field Group */
  heading?: Maybe<Scalars['String']['output']>;
  /** Field of the &quot;text&quot; Field Type added to the schema as part of the &quot;PageBuilderSectionsBeforefooterctaLayout&quot; Field Group */
  talkToTeamLabel?: Maybe<Scalars['String']['output']>;
  /** Field of the &quot;url&quot; Field Type added to the schema as part of the &quot;PageBuilderSectionsBeforefooterctaLayout&quot; Field Group */
  talkToTeamLink?: Maybe<Scalars['String']['output']>;
  /** Field of the &quot;text&quot; Field Type added to the schema as part of the &quot;PageBuilderSectionsBeforefooterctaLayout&quot; Field Group */
  tryALiveDemo?: Maybe<Scalars['String']['output']>;
};
// ...
export type PageBuilderSectionsCloudCommercePlatformLayout = AcfFieldGroup & AcfFieldGroupFields & PageBuilderSectionsCloudCommercePlatformLayout_Fields & PageBuilderSections_Layout & {
  __typename?: 'PageBuilderSectionsCloudCommercePlatformLayout';
  /** Field of the &quot;text&quot; Field Type added to the schema as part of the &quot;PageBuilderSectionsCloudCommercePlatformLayout&quot; Field Group */
  description?: Maybe<Scalars['String']['output']>;
  /**
   * The name of the field group
   * @deprecated Use __typename instead
   */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
  /** Field of the &quot;text&quot; Field Type added to the schema as part of the &quot;PageBuilderSectionsCloudCommercePlatformLayout&quot; Field Group */
  heading?: Maybe<Scalars['String']['output']>;
  /** Field of the &quot;text&quot; Field Type added to the schema as part of the &quot;PageBuilderSectionsCloudCommercePlatformLayout&quot; Field Group */
  highlightedHeading?: Maybe<Scalars['String']['output']>;
  /** Field of the &quot;image&quot; Field Type added to the schema as part of the &quot;PageBuilderSectionsCloudCommercePlatformLayout&quot; Field Group */
  icon?: Maybe<AcfMediaItemConnectionEdge>;
  /** Field of the &quot;text&quot; Field Type added to the schema as part of the &quot;PageBuilderSectionsCloudCommercePlatformLayout&quot; Field Group */
  requestADemo?: Maybe<Scalars['String']['output']>;
  /** Field of the &quot;url&quot; Field Type added to the schema as part of the &quot;PageBuilderSectionsCloudCommercePlatformLayout&quot; Field Group */
  requestADemoLink?: Maybe<Scalars['String']['output']>;
  /** Field of the &quot;image&quot; Field Type added to the schema as part of the &quot;PageBuilderSectionsCloudCommercePlatformLayout&quot; Field Group */
  sectionImage?: Maybe<AcfMediaItemConnectionEdge>;
};
// ...
export type PageBuilderSectionsCloudInnovatorsLayout = AcfFieldGroup & AcfFieldGroupFields & PageBuilderSectionsCloudInnovatorsLayout_Fields & PageBuilderSections_Layout & {
  __typename?: 'PageBuilderSectionsCloudInnovatorsLayout';
  /** Field of the &quot;repeater&quot; Field Type added to the schema as part of the &quot;PageBuilderSectionsCloudInnovatorsLayout&quot; Field Group */
  brandLogos?: Maybe<Array<Maybe<PageBuilderSectionsBrandLogos>>>;
  /**
   * The name of the field group
   * @deprecated Use __typename instead
   */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
  /** Field of the &quot;text&quot; Field Type added to the schema as part of the &quot;PageBuilderSectionsCloudInnovatorsLayout&quot; Field Group */
  headinSufix?: Maybe<Scalars['String']['output']>;
  /** Field of the &quot;text&quot; Field Type added to the schema as part of the &quot;PageBuilderSectionsCloudInnovatorsLayout&quot; Field Group */
  headingPrefix?: Maybe<Scalars['String']['output']>;
  /** Field of the &quot;text&quot; Field Type added to the schema as part of the &quot;PageBuilderSectionsCloudInnovatorsLayout&quot; Field Group */
  sectionHeading?: Maybe<Scalars['String']['output']>;
};
// ...
export type PageBuilderSectionsCloudIntegrationsPostsLayout = AcfFieldGroup & AcfFieldGroupFields & PageBuilderSectionsCloudIntegrationsPostsLayout_Fields & PageBuilderSections_Layout & {
  __typename?: 'PageBuilderSectionsCloudIntegrationsPostsLayout';
  /**
   * The name of the field group
   * @deprecated Use __typename instead
   */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
  /** Field of the &quot;text&quot; Field Type added to the schema as part of the &quot;PageBuilderSectionsCloudIntegrationsPostsLayout&quot; Field Group */
  heading?: Maybe<Scalars['String']['output']>;
  /** Field of the &quot;post_object&quot; Field Type added to the schema as part of the &quot;PageBuilderSectionsCloudIntegrationsPostsLayout&quot; Field Group */
  posts?: Maybe<AcfContentNodeConnection>;
};
// ...
export type RootQuery = {
  __typename?: 'RootQuery';
// ...
  page?: Maybe<Page>;
// ...
};
// ...
export type CloudInnovatorsSectionFragment = { __typename?: 'PageBuilderSectionsCloudInnovatorsLayout', fieldGroupName?: string | null, headinSufix?: string | null, headingPrefix?: string | null, sectionHeading?: string | null } & { ' $fragmentName'?: 'CloudInnovatorsSectionFragment' };
// ...
export type BeforeFooterCtaSectionFragment = { __typename?: 'PageBuilderSectionsBeforefooterctaLayout', fieldGroupName?: string | null, heading?: string | null, description?: string | null, talkToTeamLabel?: string | null, talkToTeamLink?: string | null, tryALiveDemo?: string | null } & { ' $fragmentName'?: 'BeforeFooterCtaSectionFragment' };

export type CloudCommercePlatformSectionFragment = { __typename?: 'PageBuilderSectionsCloudCommercePlatformLayout', fieldGroupName?: string | null, heading?: string | null, highlightedHeading?: string | null, description?: string | null, requestADemo?: string | null, requestADemoLink?: string | null, icon?: { __typename?: 'AcfMediaItemConnectionEdge', node: { __typename?: 'MediaItem', altText?: string | null, link?: string | null } } | null, sectionImage?: { __typename?: 'AcfMediaItemConnectionEdge', node: { __typename?: 'MediaItem', altText?: string | null, link?: string | null } } | null } & { ' $fragmentName'?: 'CloudCommercePlatformSectionFragment' };
// ...
export type ExperienceStackSectionFragment = { __typename?: 'PageBuilderSectionsExperienceStackLayout', fieldGroupName?: string | null } & { ' $fragmentName'?: 'ExperienceStackSectionFragment' };

export type WhyChooseStackConsoleSectionFragment = { __typename?: 'PageBuilderSectionsWhyChooseStackConsoleLayout', fieldGroupName?: string | null } & { ' $fragmentName'?: 'WhyChooseStackConsoleSectionFragment' };

export type CloudIntegrationsHeroSectionFragment = { __typename?: 'PageBuilderSectionsCloudIntegrationsHeroLayout', fieldGroupName?: string | null, heading?: string | null, headingHighlighted?: string | null, integrationButtonLabel?: string | null, integrationButtonLink?: string | null, description?: string | null, backgroundImage?: { __typename?: 'AcfMediaItemConnectionEdge', node: { __typename?: 'MediaItem', altText?: string | null, link?: string | null } } | null, brandLogos?: Array<{ __typename?: 'PageBuilderSectionsBrandLogos', fieldGroupName?: string | null, logo?: { __typename?: 'AcfMediaItemConnectionEdge', node: { __typename?: 'MediaItem', altText?: string | null, link?: string | null } } | null } | null> | null } & { ' $fragmentName'?: 'CloudIntegrationsHeroSectionFragment' };
// ...
export type PageSectionsByUriQuery = { __typename?: 'RootQuery', page?: { __typename?: 'Page', id: string, title?: string | null, slug?: string | null, uri?: string | null, featuredImageId?: string | null, link?: string | null, pageBuilder?: { __typename?: 'PageBuilder', sections?: Array<(
        { __typename: 'PageBuilderSectionsBeforefooterctaLayout', fieldGroupName?: string | null }
        & { ' $fragmentRefs'?: { 'BeforeFooterCtaSectionFragment': BeforeFooterCtaSectionFragment } }
      ) | (
        { __typename: 'PageBuilderSectionsCloudCommercePlatformLayout', fieldGroupName?: string | null }
        & { ' $fragmentRefs'?: { 'CloudCommercePlatformSectionFragment': CloudCommercePlatformSectionFragment } }
      ) | (
        { __typename: 'PageBuilderSectionsCloudInnovatorsLayout', fieldGroupName?: string | null }
        & { ' $fragmentRefs'?: { 'CloudInnovatorsSectionFragment': CloudInnovatorsSectionFragment } }
      ) | (
        { __typename: 'PageBuilderSectionsCloudIntegrationsHeroLayout', fieldGroupName?: string | null }
        & { ' $fragmentRefs'?: { 'CloudIntegrationsHeroSectionFragment': CloudIntegrationsHeroSectionFragment } }
      ) | (
        { __typename: 'PageBuilderSectionsCloudIntegrationsPostsLayout', fieldGroupName?: string | null }
        & { ' $fragmentRefs'?: { 'CloudIntegrationsPostsSectionFragment': CloudIntegrationsPostsSectionFragment } }
      ) | (
        { __typename: 'PageBuilderSectionsCloudPowerBusinessLayout', fieldGroupName?: string | null }
        & { ' $fragmentRefs'?: { 'CloudPowerBusinessSectionFragment': CloudPowerBusinessSectionFragment } }
      ) | (
        { __typename: 'PageBuilderSectionsExperienceStackLayout', fieldGroupName?: string | null }
        & { ' $fragmentRefs'?: { 'ExperienceStackSectionFragment': ExperienceStackSectionFragment } }
      ) | (
        { __typename: 'PageBuilderSectionsFaqsLayout', fieldGroupName?: string | null }
        & { ' $fragmentRefs'?: { 'FaqsSectionFragment': FaqsSectionFragment } }
      ) | (
        { __typename: 'PageBuilderSectionsHeroLayout', fieldGroupName?: string | null }
        & { ' $fragmentRefs'?: { 'HeroSectionFragment': HeroSectionFragment } }
      ) | (
        { __typename: 'PageBuilderSectionsInfiniteCloudConsoleLayout', fieldGroupName?: string | null }
        & { ' $fragmentRefs'?: { 'InfiniteCloudConsoleSectionFragment': InfiniteCloudConsoleSectionFragment } }
      ) | (
        { __typename: 'PageBuilderSectionsIntegratedCloudEcosystemLayout', fieldGroupName?: string | null }
        & { ' $fragmentRefs'?: { 'IntegratedCloudEcosystemSectionFragment': IntegratedCloudEcosystemSectionFragment } }
      ) | (
        { __typename: 'PageBuilderSectionsPlatformCapabilitiesGridLayout', fieldGroupName?: string | null }
        & { ' $fragmentRefs'?: { 'PlatformCapabilitiesGridSectionFragment': PlatformCapabilitiesGridSectionFragment } }
      ) | (
        { __typename: 'PageBuilderSectionsStackConsoleFeaturesLayout', fieldGroupName?: string | null }
        & { ' $fragmentRefs'?: { 'StackConsoleFeaturesSectionFragment': StackConsoleFeaturesSectionFragment } }
      ) | (
        { __typename: 'PageBuilderSectionsTurningideascloudpowerbusinessLayout', fieldGroupName?: string | null }
        & { ' $fragmentRefs'?: { 'TurningIdeasCloudPowerBusinessSectionFragment': TurningIdeasCloudPowerBusinessSectionFragment } }
      ) | (
        { __typename: 'PageBuilderSectionsWhoIsForStackConsoleLayout', fieldGroupName?: string | null }
        & { ' $fragmentRefs'?: { 'WhoIsForStackConsoleSectionFragment': WhoIsForStackConsoleSectionFragment } }
      ) | (
        { __typename: 'PageBuilderSectionsWhyChooseStackConsoleLayout', fieldGroupName?: string | null }
        & { ' $fragmentRefs'?: { 'WhyChooseStackConsoleSectionFragment': WhyChooseStackConsoleSectionFragment } }
      ) | null> | null } | null } | null };